Adds plant information display to UI
Implements the ability to display the plant's name and days growing on the UI.

Loads plant information from the config file and calculates the number of days since planting.
Adds console commands to set and display plant information.
Adjusts display layout to remove white bar at bottom of the screen.